Removed value prop from SearchBar
Since the name of this SearchBar prop on App.js was "value", I figured it should be called for on SearchBar.js as "props.value". But still, on SearchBar.js, it was calling it as "props.term" and it still worked. This way, I figured this "value" prop was not needed, and removing it really made no difference, continuing to work still.